Refactor components into controllers for improved organization and maintainability
