+++ lib/src/controllers/help_dialog.dart
